requirements.txt
numpy
pandas
psycopg2-binary
sqlalchemy
google-genai
python-dotenv
matplotlib
seaborn
openpyxl
tabulate
scikit-learn
xgboost
